services/example-otel/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# OpenTelemetry
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-prometheus==0.42b0

# Metrics and monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Logging
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
